Refactored to move submitEvent and submitEventInstance logic into their own components
